+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.1] - 2025-02-06
+
 ### Fixed
 
 - Fix bucket creation in us-east-1.
@@ -149,7 +151,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement creation of S3 buckets on CAPA management clusters.
 
-[Unreleased]: https://github.com/giantswarm/object-storage-operator/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/giantswarm/object-storage-operator/compare/v0.10.1...HEAD
+[0.10.1]: https://github.com/giantswarm/object-storage-operator/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/giantswarm/object-storage-operator/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/giantswarm/object-storage-operator/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/giantswarm/object-storage-operator/compare/v0.7.0...v0.8.0